+++ global_planning_algos/src/global_planning_node.cpp
@@ -124,8 +124,8 @@ int main(int argc, char** argv){
     int k_nearest;
     double robot_radius;
     int sample_size;
-    vector<double> map_x_lims;
-    vector<double> map_y_lims;
+    vector<int> map_x_lims;
+    vector<int> map_y_lims;
     double cell_size;
     nh2.getParam("obstacles", obstacle_list);
     nh2.getParam("k_nearest", k_nearest);
