+++ global_planning_algos/src/theta_star.cpp
@@ -13,8 +13,8 @@ using PRM_Grid::PRM;
 using std::vector;
 
 Theta_Star::Theta_Star(std::vector<double> start, std::vector<double> goal, XmlRpc::XmlRpcValue &obstacle_list,
-                       double cell_size, double robot_radius, int sample_size, const std::vector<double> &map_x_lims,
-                       const std::vector<double> &map_y_lims, int k_nearest) :
+                       double cell_size, double robot_radius, int sample_size, const std::vector<int> &map_x_lims,
+                       const std::vector<int> &map_y_lims, int k_nearest) :
                        A_Star(start, goal, obstacle_list, cell_size, robot_radius, sample_size, map_x_lims, map_y_lims, k_nearest)
                        {
 }
